requirements.txt
Django==2.0.3
django-cors-headers==2.2.0
djangorestframework==3.7.7
gunicorn==19.9.0
psycopg2==2.7.7
pytz==2018.9
